[nt] Start salvagers with a jetpack
With the recent randomized salvage spawn change, it is really hard to do
your job as salvage without jetpacks. Give salvagers jetpacks.

As a slight nerf, remove their extra O2 and N2 tanks, so that they'll
most likely use the internals built-in to their jetpacks. This makes
running out of air mid-salvage a real threat.