Minor resprite on xenoegg
I wasn't entirely happy with the end result of them at a closer inspection, i've done a bit more shading and reformed it slightly. i will revise this one at a later stage as well but now it's a bit close to the burrower sprite in terms of colours and shading.