Removes unnecessary message for defibbing a patient which has ghosted (#20690)
* Removes unnecessary message for defibbing a patient which has ghosted

This message is for the defib operator.

It could inform them that the patient doesn't have a consciousness since it's ghosted
but maybe no message is really the right move here because it may have succeeded
and that would be what would happen had the defib attempt succeeded - no message.

Also when the patient's ghost return to body there they will be alive as long as the
body is still alive, so telling the operator to defib again may be wrong.

* Removes corresponding ftl message which doesn't occur now