Add ignoreMaterialWhitelist
Allow lathes to ignore material whitelists. This is for furnaces to
accept any type of ore, regardless of whether or not it has a recipe to
make something using said ore.

This prevents piles of unprocessed rock from sitting in front of
furnaces. It also prevents the old blast furnace hack that uses the
blast furnace to filter ores, without you having to sort them yourself.