Moved bear hide to neck slot
Made it so that the bear hide is now worn on the neck slot. Side profile sprite a bit weird still, but I'm not certain if that's an issue with the neck layers, or something else.
Minor changes to C# code so that neck items also have the option to hide hair with the "HideHair" tag.
Bear hide now also gives a 5% protection to brute damage. Could potentially add minor protection as well, if it fits.