Add respawns and auto cloner
Brings respawn functionality via an auto cloning pod.

When a player dies or /ghosts a timer will start from the moment of death/ghosting. When the timer expires, they may respawn in an auto cloner if one is available.

The auto cloner is designed to be an indestructible, immovable piece that works under all conditions. Multiple auto-cloners can be mapped, but the first spawned cloner will be the one used if not already occupied (mutliple cloners can be used at once).

If there are no auto-cloners map or all auto-cloners are in use, the button will be deactivated and a message reading "waiting for spawn" will appear. Once an auto-cloner is available, the player can click the button to spawn as the character they selected in the lobby.

Players can also play ghost roles and still utilise the auto-cloner, spawning as their originally selected character.

The auto cloner does not need biomass and never fails - otherwise it will operate as the normal cloner does. They player will have no equipment on them once cloned, so if mapped the auto cloner should be placed in proximity to non-access controlled doors and clothing vendors.